Use edge runtime for all pages
