+++ packages/trpc/src/edgeIndex.ts
@@ -1,0 +1,1 @@
+export { getLatestExtension } from './services/extensionService';
