+++ packages/trpc/src/index.ts
@@ -3,4 +3,4 @@ export type { ILogRequest } from './constants/logs';
 export { createTRPCContext } from './trpc';
 export * from './routers';
 
-export { backupData } from './services/extensionService';
+export { backupData } from './services/backupService';
